Remove deprecated method of reading metadata
Metadata was previously stored in a separate folder inside the data
folder, and had to be read separately. With the newest version, all
metadata is contained within the .ttl-file that contains the intent,
and it is no longer necessary to have code that reads the separately.
The deprecated code was already commented out, but in this commit it is
removed completely.