Add support for Anthropic API and verify output
Support is added for Anthtopic's API.

In addition, I've added two small verifications of the agents' output
regarding intent ID and intent description. When parsing the output from
the agents' interaction, the JSON is found and parsed, and the generated
intent ID and description is checked against the original ones that were
defined in the structured intent input (.ttl file). A warning is given
if these values do not match, but the original values for the ID and
description is in any case used when saving the output, meaning the
generated ones will be overwritten. The idea is that if the generated
values do not match, it can be a warning sign that the agents have lost
context, and other details may be inaccurate as well.